+++ tests/test_api.py
@@ -4,6 +4,8 @@ import Pyro5.client
 import Pyro5.server
 import Pyro5.nameserver
 import Pyro5.callcontext
+import Pyro5.serializers
+from Pyro5.serializers import SerializerBase
 
 
 def test_api():
@@ -14,3 +16,7 @@ def test_api():
     assert Pyro5.api.Daemon is Pyro5.server.Daemon
     assert Pyro5.api.start_ns is Pyro5.nameserver.start_ns
     assert Pyro5.api.current_context is Pyro5.callcontext.current_context
+    assert Pyro5.api.register_dict_to_class == SerializerBase.register_dict_to_class
+    assert Pyro5.api.register_class_to_dict == SerializerBase.register_class_to_dict
+    assert Pyro5.api.unregister_dict_to_class == SerializerBase.unregister_dict_to_class
+    assert Pyro5.api.unregister_class_to_dict == SerializerBase.unregister_class_to_dict